feat: wire JIT dispatch + fix P2 perf bugs + add e2e JIT tests
JIT Dispatch Wiring (the missing link):
- Add JIT dispatch check in interpreter CALL_STATIC handler:
  when target_method->compiled_code != NULL, call JIT code
  directly via function pointer instead of creating interpreter frame
- Add vtx_dispatch_jit() helper with atomic load of compiled_code
  and fallback to interpreter on deopt
- Fix method_id hardcode (was 0): now uses vtable_index or
  method pointer hash
- Add emit_test_reg_reg2 for two-register TEST instruction
- Replace IMUL overflow guard (deopt) with inline JO fallback
  path that produces SMI(0) instead of crashing on missing deopt_info

P2 Performance Bug Fixes:
- Clock eviction use_bit race: use atomic store/load
- Poly IC memory leak: track in compiled_method_t, free on eviction
- Code cache total_size alignment: match 16-byte alignment on free
- Code buffer infinite loop guard when capacity=0
- Profiler OOM: LRU eviction fallback instead of silent data loss
- Compilation_requested array race: use atomic operations
- Mid-tier stats: write to global instead of local variable
- compilation_time_ns: change uint32_t to uint64_t

End-to-End JIT Tests (5/7 pass):
- jit_identity, jit_const_return, jit_simple_add, jit_simple_sub,
  jit_bitwise_and all pass
- jit_simple_mul: IMUL fast path has overflow detection bug (returns 0)
- jit_comparison_branch: untag_smi sign-extension bug for negative values

Known codegen bugs identified by e2e testing:
1. IMUL fast path: JO incorrectly fires for non-overflowing multiplies
2. SMI untag sign-extension: negative SMI values lose sign after
   AND mask + sign-extend sequence, causing wrong icmp_gt results

+++ src/midtier/midtier.h
@@ -35,11 +35,11 @@ typedef struct {
     uint32_t inline_max_size;   /* max callee bytecode size for inlining (0 = no inlining) */
     bool specialize_arithmetic; /* specialize integer/float ops based on type feedback? */
 
-    /* Statistics */
+    /* Statistics — stored in global struct so they persist across compilations */
     uint32_t methods_compiled;
     uint32_t sites_specialized;
     uint32_t guards_inserted;
-    uint32_t compilation_time_ns;
+    uint64_t compilation_time_ns;  /* total compilation time in nanoseconds (uint64 to avoid wrap) */
 } vtx_midtier_config_t;
 
 /* Result of mid-tier compilation */
